# src/internal/http_api/swagger/swagger.yaml
definitions:
  models.AddTeamRequest:
    properties:
      members:
        items:
          $ref: '#/definitions/models.TeamMemberRequest'
        type: array
      team_name:
        type: string
    required:
    - members
    - team_name
    type: object
  models.AddTeamResponse:
    properties:
      team:
        $ref: '#/definitions/models.TeamResponse'
    type: object
  models.CreatePullRequestRequest:
    properties:
      author_id:
        type: string
      pull_request_id:
        type: string
      pull_request_name:
        type: string
    required:
    - author_id
    - pull_request_id
    - pull_request_name
    type: object
  models.ErrorBody:
    properties:
      code:
        $ref: '#/definitions/models.ErrorCode'
      message:
        type: string
    type: object
  models.ErrorCode:
    enum:
    - TEAM_EXISTS
    - PR_EXISTS
    - PR_MERGED
    - NOT_ASSIGNED
    - NO_CANDIDATE
    - NOT_FOUND
    - DECODE_FAILED
    - VALIDATION_FAILED
    - INTERNAL_SERVER_ERROR
    type: string
    x-enum-varnames:
    - ErrorCodeTeamExists
    - ErrorCodePRExists
    - ErrorCodePRMerged
    - ErrorCodeNotAssigned
    - ErrorCodeNoCandidate
    - ErrorCodeNotFound
    - ErrorCodeDecodeFailed
    - ErrorCodeValidationFailed
    - ErrorCodeInternalServer
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.ErrorBody'
    type: object
  models.GetUserReviewsResponse:
    properties:
      pull_requests:
        items:
          $ref: '#/definitions/models.PullRequestShortResponse'
        type: array
      user_id:
        type: string
    type: object
  models.HealthResponse:
    properties:
      status:
        type: string
    type: object
  models.MergePullRequestRequest:
    properties:
      pull_request_id:
        type: string
    required:
    - pull_request_id
    type: object
  models.PullRequestEnvelopeResponse:
    properties:
      pr:
        $ref: '#/definitions/models.PullRequestResponse'
    type: object
  models.PullRequestResponse:
    properties:
      assigned_reviewers:
        items:
          type: string
        type: array
      author_id:
        type: string
      createdAt:
        type: string
      mergedAt:
        type: string
      pull_request_id:
        type: string
      pull_request_name:
        type: string
      status:
        type: string
    type: object
  models.PullRequestShortResponse:
    properties:
      author_id:
        type: string
      pull_request_id:
        type: string
      pull_request_name:
        type: string
      status:
        type: string
    type: object
  models.ReassignPullRequestRequest:
    properties:
      old_reviewer_id:
        type: string
      pull_request_id:
        type: string
    required:
    - old_reviewer_id
    - pull_request_id
    type: object
  models.ReassignPullRequestResponse:
    properties:
      pr:
        $ref: '#/definitions/models.PullRequestResponse'
      replaced_by:
        type: string
    type: object
  models.SetUserIsActiveRequest:
    properties:
      is_active:
        type: boolean
      user_id:
        type: string
    required:
    - user_id
    type: object
  models.SetUserIsActiveResponse:
    properties:
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.TeamMemberRequest:
    properties:
      is_active:
        type: boolean
      user_id:
        type: string
      username:
        type: string
    required:
    - user_id
    - username
    type: object
  models.TeamMemberResponse:
    properties:
      is_active:
        type: boolean
      user_id:
        type: string
      username:
        type: string
    type: object
  models.TeamResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/models.TeamMemberResponse'
        type: array
      team_name:
        type: string
    type: object
  models.TeamStatsResponse:
    properties:
      active_members_count:
        type: integer
      avg_time_to_merge_seconds:
        type: integer
      members_count:
        type: integer
      merged_prs:
        type: integer
      open_prs:
        type: integer
      team_name:
        type: string
      total_prs:
        type: integer
    type: object
  models.UserResponse:
    properties:
      is_active:
        type: boolean
      team_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  title: PR Reviewer Assignment Service (Test Task, Fall 2025)
  version: 1.0.0
paths:
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HealthResponse'
      summary: Health check
      tags:
      - Health
  /pullRequest/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create pull request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreatePullRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: PR создан
          schema:
            $ref: '#/definitions/models.PullRequestEnvelopeResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Автор/команда не найдены
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: PR уже существует
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать PR и автоматически назначить до 2 ревьюверов из команды автора
      tags:
      - PullRequests
  /pullRequest/merge:
    post:
      consumes:
      - application/json
      parameters:
      - description: Merge pull request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MergePullRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PR в состоянии MERGED
          schema:
            $ref: '#/definitions/models.PullRequestEnvelopeResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: PR не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Пометить PR как MERGED (идемпотентная операция)
      tags:
      - PullRequests
  /pullRequest/reassign:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reassign pull request reviewer body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ReassignPullRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Переназначение выполнено
          schema:
            $ref: '#/definitions/models.ReassignPullRequestResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: PR или пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Нарушение доменных правил переназначения
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Переназначить конкретного ревьювера на другого из его команды
      tags:
      - PullRequests
  /team/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Add team body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Команда создана
          schema:
            $ref: '#/definitions/models.AddTeamResponse'
        "400":
          description: Команда уже существует или неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Создать команду с участниками (создаёт/обновляет пользователей)
      tags:
      - Teams
  /team/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: Уникальное имя команды
        in: query
        name: team_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Объект команды
          schema:
            $ref: '#/definitions/models.TeamResponse'
        "404":
          description: Команда не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить команду с участниками
      tags:
      - Teams
  /team/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: Уникальное имя команды
        in: query
        name: team_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статистика по команде
          schema:
            $ref: '#/definitions/models.TeamStatsResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Команда не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить статистику по команде
      tags:
      - Teams
  /users/getReview:
    get:
      consumes:
      - application/json
      parameters:
      - description: Идентификатор пользователя
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список PR'ов пользователя
          schema:
            $ref: '#/definitions/models.GetUserReviewsResponse'
        "400":
          description: отсутствующий или неверный user_id
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить PR'ы, где пользователь назначен ревьювером
      tags:
      - Users
  /users/setIsActive:
    post:
      consumes:
      - application/json
      parameters:
      - description: Set is active body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SetUserIsActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый пользователь
          schema:
            $ref: '#/definitions/models.SetUserIsActiveResponse'
        "400":
          description: неверный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Установить флаг активности пользователя
      tags:
      - Users
swagger: "2.0"